Add 'Secondary Fire Powerup' feature.
This is an enemy seeking missile that tracks the closest enemy to the missile. When the missile spawns it finds the nearest enemy to it, but if the enemy respawns it will refind the nearest enemy to it and track that one.

Some extra changes that went into this one: 1) ammo count is not decremented when using powerups because then weapon powerups can't be used when ammo count is 0. 2) Various refactorings to support this feature, and 3) sprite and audio packs to support the feature.