Added feature "New Enemy Type" (#21)
New enemy is an enemy that zig-zags left and right as it goes downward. It has an AOE bomb-weapon (instead of double lasers) that when exploded will damage player. Collision with the bomb will cause AOE to explode and damage player. Players re-entering the AOE after escaping it will take more damage.